@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.1.4+14
+version: 0.1.5+15
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=3.0.6 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,12 +33,13 @@ dependencies:
     git:
       url: https://github.com/candidelabs/wallet-dart.git
       ref: 247af6f807c259ff85e09803250e9ccb0889d8cd
-  ens_dart: ^1.0.0
   magic_sdk: ^2.0.2
   dio: ^5.1.2
+  web_socket_channel: ^2.4.0
+  http: ^0.13.5
   eth_sig_util: ^0.0.9
   walletconnect_dart: ^0.0.11
-  walletconnect_flutter_v2: ^2.0.12
+  walletconnect_flutter_v2: ^2.1.3
   # Auth
   biometric_storage: ^5.0.0
   # State
