@@ -1,7 +1,6 @@
 import 'package:candide_mobile_app/config/network.dart';
 import 'package:candide_mobile_app/controller/token_info_storage.dart';
-import 'package:candide_mobile_app/utils/constants.dart';
-import 'package:wallet_dart/wallet/user_operation.dart';
+import 'package:candide_mobile_app/models/gas.dart';
 
 class FeeToken {
   TokenInfo token;
@@ -13,14 +12,13 @@ class FeeToken {
 
   static final BigInt costOfPost = BigInt.from(45000); // todo shouldn't be hardcoded
 
-  BigInt calculateETHFee(UserOperation op, Network network) {
-    bool isEther = token.symbol == network.nativeCurrency && token.address == Constants.addressZeroHex;
-    BigInt operationMaxEthCost = op.maxFeePerGas * (costOfPost + op.callGasLimit + (op.verificationGasLimit * BigInt.from(isEther ? 1 : 3)) + op.preVerificationGas);
+  BigInt calculateETHFee(GasEstimate gasEstimate, Network network, bool withPaymaster) {
+    BigInt operationMaxEthCost = gasEstimate.maxFeePerGas * (costOfPost + gasEstimate.callGasLimit + (gasEstimate.verificationGasLimit * BigInt.from(!withPaymaster ? 1 : 3)) + gasEstimate.preVerificationGas);
     return operationMaxEthCost;
   }
 
-  BigInt calculateFee(UserOperation op, Network network) {
-    BigInt operationMaxEthCost = calculateETHFee(op, network);
+  BigInt calculateFee(GasEstimate gasEstimate, Network network, bool withPaymaster) {
+    BigInt operationMaxEthCost = calculateETHFee(gasEstimate, network, withPaymaster);
     BigInt tokenFee = (operationMaxEthCost * exchangeRate) ~/ BigInt.from(10).pow(18);
     tokenFee += paymasterFee;
     return tokenFee;
