@@ -5,8 +5,7 @@ import 'package:candide_mobile_app/models/paymaster/sponsor_data.dart';
 class PaymasterResponse {
     List<FeeToken> tokens;
     SponsorData sponsorData;
-    PaymasterData paymasterData;
+    PaymasterMetadata paymasterData;
 
     PaymasterResponse({required this.tokens, required this.sponsorData, required this.paymasterData});
 }
-
